+++ database/migrations/2022_01_28_191957_create_produto_descricao.php
@@ -15,7 +15,7 @@ class CreateProdutoDescricao extends Migration
     {
         Schema::create('produto_descricao', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('produto_id');
+            $table->unsignedInteger('produto_id');
             $table->string('descricao');
             $table->timestamps();
         });
