+++ database/migrations/2022_01_28_195702_create_produto_preco_especial.php
@@ -15,7 +15,7 @@ class CreateProdutoPrecoEspecial extends Migration
     {
         Schema::create('produto_preco_especial', function (Blueprint $table) {
             $table->increments('id');
-            $table->unsignedBigInteger('produto_id');
+            $table->unsignedInteger('produto_id');
             $table->decimal('preco',10,2);
             $table->date('data_inicio');
             $table->date('data_fim');
